fix ip blocklist crashing LocalServer panel
in the IP Blocklist panel, the unlock button was added at an index
position. this was not necessary, but cascaded the issue that when the
refresh button was not added at 0, inserting at index 1 caused errors
cascading upto LocalServer.

the refresh button did not need to be added before the commit separately
because it was composed by a buttonPanel.addAction method that did both
object creation and addition to the panel together before commit
c10f6027a38f554a7227b2772c9ec809a341bc4c, and after that commit was
created by a Button constructor.